requirements.txt
speechrecognition
pyttsx3
requests
dotenv
google-generativeai
pyaudio